+++ WebApplication/src/components/ResourceDetailText.vue
@@ -1,9 +1,10 @@
 <script>
 import CopyToClipboardButton from "./CopyToClipboardButton.vue";
+import { translateDicomTag } from "../locales/i18n"
 
 
 export default {
-    props: ["title", "value", "truncate", "copy"], // you can set copy=false if you don't wan't the copy button !
+    props: ["truncate", "copy", "tags", "tag", "showIfEmpty"], // you can set copy=false if you don't wan't the copy button !
     setup() {
         return {
         }
@@ -17,6 +18,22 @@ export default {
     methods: {
     },
     computed: {
+        title() {
+            return translateDicomTag(this.$i18n.t, this.tag);
+        },
+        value() {
+            if (this.hasValue) {
+                return this.tags[this.tag];
+            } else {
+                return "";
+            }
+        },
+        hasValue() {
+            return this.tag in this.tags;
+        },
+        showValue() {
+            return this.showIfEmpty || this.hasValue;
+        }
     },
     components: { CopyToClipboardButton }
 }
@@ -24,10 +41,11 @@ export default {
 
 <template>
 
-    <li class="d-flex">
-        <span class="details-label" :class="{'d-inline-block text-truncate': this.truncate}"><slot></slot>: </span>
-        <span class="details">{{ value }}</span>
-        <CopyToClipboardButton v-if="!this.copy && this.value && this.value.length > 0" :valueToCopy="this.value" />
+    <li v-if="showValue" class="d-flex">
+        <span class="details-label" :class="{ 'd-inline-block text-truncate': this.truncate }" data-bs-toggle="tooltip"
+            v-bind:title="title">{{ title }}: </span>
+        <span class="details" data-bs-toggle="tooltip" v-bind:title="value">{{ value }}</span>
+        <CopyToClipboardButton v-if="!this.copy && value && value.length > 0" :valueToCopy="value" />
     </li>
 
 </template>