+++ WebApplication/src/orthancApi.js
@@ -3,68 +3,9 @@ import store from "./store"
 
 import { orthancApiUrl, oe2ApiUrl } from "./globalConfigurations";
 
-
-// default config only used when running dev against an Orthanc which does not have the latest plugin
-const _defaultConfig = {
-    "UiOptions" : {
-        "EnableUpload": true,
-        "EnableDeleteResources": true,
-        "EnableApiViewMenu": true,
-        "EnableSettings": true,
-        "MaxStudiesDisplayed": 50,
-        "UploadReportTags" : [
-            "AccessionNumber", 
-            "StudyDate", 
-            "PatientName", 
-            "PatientID", 
-            "StudyDescription"
-        ],
-        "UploadReportMaxTags": 2,
-        "StudyListColumns" : [
-            "PatientBirthDate",
-            "PatientName",
-            "PatientID",
-            "StudyDescription",
-            "StudyDate",
-            "modalities",
-            "AccessionNumber",
-            "seriesCount"
-    ]
-    },
-    "Plugins": {
-		"dicom-web" : {
-			"Description" : "Implementation of DICOMweb (QIDO-RS, STOW-RS and WADO-RS) and WADO-URI.",
-			"Enabled" : true,
-			"ExtendsOrthancExplorer" : true,
-			"ID" : "dicom-web",
-			"RootUri" : "../dicom-web/app/client/index.html",
-			"Version" : "mainline"
-		},
-        "osimis-web-viewer": {
-            "Description" : "Provides a Web viewer of DICOM series within Orthanc.",
-            "ExtendsOrthancExplorer" : true,
-            "ID" : "osimis-web-viewer",
-            "Version" : "1.4.2.0-9d9eff4",
-			"Enabled" : false,
-        },
-		"stone-webviewer" : 
-		{
-			"Enabled" : true,
-			"ExtendsOrthancExplorer" : true,
-			"ID" : "stone-webviewer",
-			"Version" : "2.2"
-		}
-    },
-}
-
 export default {
     async loadOe2Configuration() {
-        try {
-            return (await axios.get(oe2ApiUrl + "configuration")).data;
-        } catch (e) {
-            console.error("Error while loading OE2 configuration: ", e);
-            return _defaultConfig;
-        }
+        return (await axios.get(oe2ApiUrl + "configuration")).data;
     },
     async loadDicomWebServers() {
         return (await axios.get(orthancApiUrl + "dicom-web/servers")).data;
@@ -253,6 +194,17 @@ export default {
         }));
         
         return response.data['url'];
+    },
+
+    async modifyStudy(orthancId, replaceTags, removeTags) {
+        const response = (await axios.post(orthancApiUrl + "studies/" + orthancId + "/modify", {
+            "Replace" : replaceTags,
+            "Remove" : removeTags,
+            "Force": true,
+            "Synchronous": true
+        }))
+
+        return response.data;
     },
 
     ////////////////////////////////////////// HELPERS