+++ release-notes.md
@@ -1,3 +1,10 @@
+Changes:
+- allow modification of studies
+- new configuration 'Modifications'
+- new configuration 'TranslateDicomTags'   TODO
+- new configurations 'StudyMainDicomTags', 'PatientMainDicomTags' to define the list of tags to display in the study details
+
+
 0.5.1 (2022-12-20)
 ==================